Add animated hero landing page
