+++ src/components/Navbar.tsx
@@ -108,6 +108,9 @@ export default function Navbar() {
           >
             Home
           </Link>
+          <Link href="/rankings" className={`${pathname === "/rankings" ? "underline" : "hover:underline"}`}>
+            Rankings
+          </Link>
 
           {profileUsername && (
             <Link
@@ -166,6 +169,9 @@ export default function Navbar() {
           >
             Home
           </Link>
+            <Link href="/rankings" onClick={() => setMenuOpen(false)} className="w-full text-center py-1">
+              Rankings
+            </Link>
           {profileUsername && (
             <Link
               href={`/profile/${profileUsername}`}
